backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.3.3
openpyxl==3.1.2
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1